requirements.txt
pytest
selenium
mysql-connector-python
allure-pytest
configparser